Add files via upload
This is  a python quiz game.
The player will be asked if they want to play. 
If yes , the program continues .If no, the program exits.
Ten computer acronym and abbreviation questions will be asked.
If the answer is correct, one point will be added to their score.
At the end of the questions a score in percentage will be given.